Refactor saving and dispatching process within try block
This commit moves the saving of entities and dispatching of events into the try block of their respective methods. It does this for the Active Campaign methods in multiple consumer files. By putting these functionalities within the try block, we ensure the process is more seamless and only complete when no exception is thrown.